+++ primary/joystick.h
@@ -76,9 +76,9 @@ struct Joystick {
 		res.x = maths::map(float(info.dwXpos), float(fCaps.wXmin), float(fCaps.wXmax), -1, 1);
 		res.y = maths::map(float(info.dwYpos), float(fCaps.wYmin), float(fCaps.wYmax), 1, -1);
 		res.z = maths::map(float(info.dwZpos), float(fCaps.wZmin), float(fCaps.wZmax), 1, 0); //throttle reverse
-		res.r = maths::map(float(info.dwRpos), fCaps.wRmin, fCaps.wRmax, -1, 1);
-		res.u = maths::map(float(info.dwUpos), fCaps.wUmin, fCaps.wUmax, -1, 1);
-		res.v = maths::map(float(info.dwVpos), fCaps.wVmin, fCaps.wVmax, -1, 1);
+		res.r = maths::map(float(info.dwRpos), (float)fCaps.wRmin, (float)fCaps.wRmax, -1, 1);
+		res.u = maths::map(float(info.dwUpos), (float)fCaps.wUmin, (float)fCaps.wUmax, -1, 1);
+		res.v = maths::map(float(info.dwVpos), (float)fCaps.wVmin, (float)fCaps.wVmax, -1, 1);
 		
 		res.buttons = info.dwButtons;
 		res.numButtonsPressed = info.dwButtonNumber;
